+++ packages/react/src/components/PageBanner/index.ts
@@ -1,0 +1,16 @@
+export {
+  PageBannerCloseButton,
+  PageBannerTextContent,
+  PageBannerDescription,
+  PageBannerButton,
+  PageBannerRoot,
+  PageBannerTitle,
+  type PageBannerCloseButtonProps,
+  type PageBannerTextContentProps,
+  type PageBannerDescriptionProps,
+  type PageBannerButtonProps,
+  type PageBannerRootProps,
+  type PageBannerTitleProps,
+} from "./PageBanner";
+
+export * as PageBanner from "./PageBanner.namespace";